% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_estimate.R
\name{block_estimate}
\alias{block_estimate}
\title{Calculate statistics needed for controling the disclosure}
\usage{
block_estimate(x, variable, r = 200, ..., field = variable)
}
\arguments{
\item{x}{\link[sp:SpatialPointsDataFrame]{sp::SpatialPointsDataFrame} or \link[sf:sf]{sf::sf} object that is used to create a raster.}

\item{variable}{name of data column or numeric with same length as \code{x}
to be used for the data.}

\item{r}{either a desired resolution or a pre-existing \link[raster:raster]{raster::raster} object.}

\item{...}{passed through to \code{\link[raster:rasterize]{raster::rasterize()}}}

\item{field}{synonym for \code{variable}. If both supplied, \code{field} has precedence.}
}
\description{
Calculate statistics needed for controling the disclosure
}
\examples{
library(raster)
block <- block_estimate(enterprises, field = "production", r = 500)
}
