% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_block.R
\name{calc_block}
\alias{calc_block}
\title{Calculate statistics needed for controling the disclosure}
\usage{
calc_block(x, variable, r = 200, ..., field = variable)
}
\arguments{
\item{x}{\link[sp:SpatialPointsDataFrame]{sp::SpatialPointsDataFrame} or \link[sf:sf]{sf::sf} object that is used to create a raster.}

\item{variable}{name of data column to be used for the data.}

\item{r}{either a desired resolution or a pre-existing \link[raster:raster]{raster::raster} object.}

\item{...}{passed through to \code{\link[raster:rasterize]{raster::rasterize()}}}

\item{field}{synonym for \code{variable}. If both supplied, \code{field} has precedence.}
}
\description{
Calculate statistics needed for controling the disclosure
}
