% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raster.R
\name{create_quad_raster}
\alias{create_quad_raster}
\title{Create a quad raster at a certain resolution}
\usage{
create_quad_raster(x, res = 100, ...)
}
\arguments{
\item{x}{geographical object (that supports extent).}

\item{res}{desired resolution (default 200).}

\item{...}{passed through to \code{\link{raster}} function.}
}
\description{
Utility function to generate a raster at a certain resolution.
}
