% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_at.R
\name{create_raster}
\alias{create_raster}
\title{Create a raster at a certain resolution}
\usage{
create_raster(x, res = 100, ...)
}
\arguments{
\item{x}{geographical object (that supports extent).}

\item{res}{desired resolution (default 100).}

\item{...}{passed through to \code{\link{raster}} function.}
}
\description{
Utility function to generate a raster at a certain resolution (with rounded coordinates).
}
