% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_risk.R
\name{disclosure_risk}
\alias{disclosure_risk}
\title{Calculate disclosure risk}
\usage{
disclosure_risk(x, risk_type = x$risk_type)
}
\arguments{
\item{x}{\link{sdc_raster} object created with \code{\link[=sdc_raster]{sdc_raster()}}}

\item{risk_type}{what kind of measure should be used.}
}
\value{
\link[raster:raster]{raster::raster} object with the disclosure risk.
}
\description{
Different risk functions can be used:
}
\details{
\itemize{
\item external (numeric variable), calculates how much the largest value comprises the total sum
\item internal (numeric variable), calculates how much the largest value comprises the sum without the second largest value
\item discrete (logical variable), calculates the fraction of sensitive values.
}
}
\seealso{
Other sensitive: \code{\link{is_sensitive}},
  \code{\link{remove_sensitive}}, \code{\link{sdc_raster}},
  \code{\link{sensitivity_score}}
}
\concept{sensitive}
