% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dwellings}
\alias{dwellings}
\title{Data set generated with residence locations  from the open data BAG register (reference)
The locations are realistic, but the associated data is simulated.}
\format{An object of class \code{data.frame} with 90603 rows and 2 columns.}
\usage{
dwellings
}
\description{
\itemize{
\item \code{x} x coordinate in Dutch crs (epsg 28992)
\item \code{y} y coordinate in Dutch crs (epsg 28992)
\item \code{sens_cont} simulated sensitive continuous value
\item \code{sens_discrete} simulated sensitive discrete value
}
}
\keyword{datasets}
