% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matrix.R
\name{extract_matrix}
\alias{extract_matrix}
\title{extracts the matrix from a raster, so that it conforms to geospatial data}
\usage{
extract_matrix(r)
}
\description{
extracts the matrix from a raster, so that it conforms to geospatial data
}
