% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_at.R
\name{get_raster_at}
\alias{get_raster_at}
\title{Create a raster at a certain resolution}
\usage{
get_raster_at(x, res = 100, ...)
}
\arguments{
\item{x}{geographical object (that supports extent).}

\item{res}{desired resolution (default 100).}

\item{...}{passed through to \code{\link{raster}} function.}
}
\description{
Utility function to generate a raster at a certain resolution (with rounded coordinates).
}
