% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff.R
\name{hellinger_distance}
\alias{hellinger_distance}
\title{hellinger distance}
\usage{
hellinger_distance(x, y, ..., normalize = TRUE)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}
}
\description{
Hellinger distance calculates the difference between two
probability distributions.
}
