% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_disclosed.R
\name{is_disclosed}
\alias{is_disclosed}
\title{Find out which raster cells are unsafe}
\usage{
is_disclosed(r, value, p = 0.95, internal = FALSE, ...)
}
\arguments{
\item{r}{Raster object}

\item{value}{?}

\item{p}{threshold when}

\item{...}{passed through to raster function}
}
\description{
Find out which raster cells are disclosed and are considered unsafe.
}
