% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sensitive.R
\name{is_sensitive}
\alias{is_sensitive}
\title{Raster with sensitive locations}
\usage{
is_sensitive(x, max_risk = x$max_risk, min_count = x$min_count,
  risk_type = c("external", "internal", "discrete"))
}
\arguments{
\item{x}{\link{sdc_raster} object created with \code{\link[=sdc_raster]{sdc_raster()}}}

\item{max_risk}{a risk value higher than \code{max_risk} will be sensitive.}

\item{min_count}{a count lower than \code{min_count} will be sensitive.}

\item{risk_type}{what kind of measure should be used. (Details)}
}
\description{
Create a binary raster with sensitive locations.
}
\details{
Different risk functions can be used:
\itemize{
\item external (numeric variable), calculates how much the largest value comprises the total sum
\item internal (numeric variable), calculates how much the largest value comprises the sum without the second largest value
\item discrete (logical variable), calculates the fraction of sensitive values.
}
}
\seealso{
Other sensitive: \code{\link{disclosure_risk}},
  \code{\link{remove_sensitive}},
  \code{\link{sensitivity_score}}
}
\concept{sensitive}
