% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{is_unsafe}
\alias{is_unsafe}
\title{Raster with unsafe locations}
\usage{
is_unsafe(x, max_risk = 0.95, min_count = 10, type = c("external",
  "internal", "discrete"))
}
\arguments{
\item{x}{\link{sdcmap} object created with \code{\link[=block_estimate]{block_estimate()}}}

\item{max_risk}{a risk value higher than \code{max_risk} will be unsafe.}

\item{min_count}{a count lower than \code{min_count} will be unsafe.}

\item{type}{what kind of measure should be used. (Details)}
}
\description{
Create a binary raster with unsafe locations.
}
\details{
Different risk functions can be used:
\itemize{
\item external (numeric variable), calculates how much the largest value comprises the total sum
\item internal (numeric variable), calculates how much the largest value comprises the sum without the second largest value
\item discrete (logical variable), calculates the fraction of sensitive values.
}
}
\seealso{
Other disclosed: \code{\link{disclosure_risk}},
  \code{\link{remove_unsafe}}
}
\concept{disclosed}
