% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_rel_freq.R
\name{kde2_rel_freq}
\alias{kde2_rel_freq}
\title{bkde2D estimate of relative frequency (indirect quotient)}
\usage{
kde2_rel_freq(spdf, width = 100, Nc = "c", max_freq = 0.9,
  resolution = 50, k0 = 10)
}
\arguments{
\item{width}{Spatial kernel density estimation band width}

\item{Nc}{name of column with classified property}

\item{max_freq}{maximum allowed frequency. Higher frequency will be cut off.}

\item{resolution}{of resulting raster}

\item{k0}{minimum number of elements}

\item{spfd}{SpatialPointsDataFrame with "Nc"}
}
\description{
Make kde plot of a spatial points data.frame
}
