% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_grid.R
\name{mask_grid}
\alias{mask_grid}
\title{Mask coordinates using a grid}
\usage{
mask_grid(x, r)
}
\description{
Pertubates coordinates by rounding coordinates to grid coordinates
}
\seealso{
Other point pertubation: 
\code{\link{mask_random}()},
\code{\link{mask_voronoi}()},
\code{\link{mask_weighted_random}()}
}
\concept{point pertubation}
