% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_random.R
\name{mask_random}
\alias{mask_random}
\title{Mask coordinates using random pertubation}
\usage{
mask_random(x, r)
}
\arguments{
\item{x}{coordinates, \code{matrix} or \code{data.frame} (first two columns)}

\item{r}{\code{numeric} maximum pertubation distance (vectorized)}
}
\value{
adapted \code{x} with perturbed coordinates
}
\description{
Pertubates points with a uniform pertubation in a circle.
Note that \code{r} can either be one distance, of a distance per data point.
}
\seealso{
Other point pertubation: 
\code{\link{mask_grid}()},
\code{\link{mask_voronoi}()},
\code{\link{mask_weighted_random}()}
}
\concept{point pertubation}
