% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_weighted_random.R
\name{mask_weighted_random}
\alias{mask_weighted_random}
\title{Mask coordinates using weighted random pertubation}
\usage{
mask_weighted_random(x, k = 5, r = NULL)
}
\arguments{
\item{x}{coordinates, \code{matrix} or \code{data.frame} (first two columns)}

\item{k}{\code{integer} number of neighbors to be used as the maximum distance}

\item{r}{\code{numeric} maximum pertubation distance (vectorized)}
}
\description{
This method uses per point the distance to the \code{k}th neighbor as the maximum
pertubation distance. Parameter \code{r} can be used to restrict the maximum distance
of the \code{k}th neighbor.
}
\references{
Spatial obfuscation methods for privacy protection of
household-level data
}
\seealso{
Other point pertubation: 
\code{\link{mask_grid}()},
\code{\link{mask_random}()},
\code{\link{mask_voronoi}()}
}
\concept{point pertubation}
