% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_values.R
\name{masked_values}
\alias{masked_values}
\title{Mask coordinates using random pertubation}
\usage{
masked_values(x, r, plot = FALSE)
}
\arguments{
\item{x}{coordinates, \code{matrix} or \code{data.frame} (first two columns)}

\item{r}{\code{\link[=sdc_raster]{sdc_raster()}} used to generate a value for the observation at hand}
}
\value{
data.frame \code{x} with (added) masked_value column
}
\description{
Pertubates points with a uniform pertubation in a circle.
Note that \code{r} can either be one distance, of a distance per data point.
}
\examples{
x <- cbind(
  x = c(2.5, 3.5, 7.2, 1.5),
  y = c(6.2, 3.8, 4.4, 2.1)
)

# plotting is only useful from small datasets!

# grid masking
x_g <- mask_grid(x, r=1, plot=TRUE)

# random pertubation
set.seed(3)
x_r <- mask_random(x, r=1, plot=TRUE)

if (requireNamespace("FNN", quietly = TRUE)){
  # weighted random pertubation
  x_wr <- mask_weighted_random(x, k = 2, r = 4, plot=TRUE)
}

if ( requireNamespace("FNN", quietly = TRUE)
  && requireNamespace("sf", quietly = TRUE)
   ){
  # voronoi masking, plotting needs package `sf`
  x_vor <- mask_voronoi(x, r = 1, plot=TRUE)
}

r_smooth <-
  sdc_raster(enterprises, "production") |>
  protect_smooth() |>
  remove_sensitive()

mask_values(enterprises, r_smooth)
}
