% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sensitive}
\alias{plot_sensitive}
\title{Plot the sensitive cells of the sdc_raster}
\usage{
plot_sensitive(x, fun = mean, ...)
}
\arguments{
\item{x}{\link{sdc_raster} object}

\item{fun}{function  used to create visualisation.}

\item{...}{Arguments passed on to \code{is_sensitive}
\describe{
  \item{x}{\link{sdc_raster} object created with \code{\link[=sdc_raster]{sdc_raster()}}}
  \item{max_risk}{a risk value higher than \code{max_risk} will be sensitive.}
  \item{min_count}{a count lower than \code{min_count} will be sensitive.}
  \item{risk_type}{what kind of measure should be used. (Details)}
}}
}
\description{
Plot the sensitive cells of the sdc_raster
}
\seealso{
Other plotting: \code{\link{plot.sdc_raster}}
}
\concept{plotting}
