% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_quadtree.R
\name{protect_quadtree}
\alias{protect_quadtree}
\title{quadtree method}
\usage{
protect_quadtree(r, ...)
}
\arguments{
\item{x}{\code{\link{sdc_raster}} object}
}
\value{
a \code{\link{sdc_raster}} object, which has been
}
\description{
quadtree method
}
\keyword{internal}
