% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sensitive.R
\name{remove_sensitive}
\alias{remove_sensitive}
\alias{mask_sensitive}
\title{remove sensitive cells from raster}
\usage{
remove_sensitive(x, max_risk = x$max_risk, min_count = x$min_count,
  ...)

mask_sensitive(x, max_risk = x$max_risk, min_count = x$min_count, ...)
}
\arguments{
\item{x}{\link{sdc_raster} object created with \code{\link[=sdc_raster]{sdc_raster()}}}

\item{max_risk}{a risk value higher than \code{max_risk} will be sensitive.}

\item{min_count}{a count lower than \code{min_count} will be sensitive.}

\item{...}{passed on to \code{\link{is_sensitive}}.}
}
\value{
sdc_raster object with sensitive cells set to \code{NA}.
}
\description{
remove sensitive cells from raster
}
\examples{
library(raster)

b <- sdc_raster(dwellings, "unemployed", r=200)

# plot the normally rastered data
plot(b, zlim=c(0,1))
plot_sensitive(b)

b_safe <- remove_sensitive(b, risk_type="discrete")
plot(b_safe, zlim=c(0,1))
}
\seealso{
Other disclosed: \code{\link{disclosure_risk}},
  \code{\link{is_sensitive}}
}
\concept{disclosed}
