% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unsafe.R
\name{remove_unsafe}
\alias{remove_unsafe}
\title{remove unsafe cells from raster}
\usage{
remove_unsafe(x, max_risk = 0.95, min_count = 10, ...)
}
\arguments{
\item{x}{\link{sdcmap} object created with \code{\link[=block_estimate]{block_estimate()}}}

\item{max_risk}{a risk value higher than \code{max_risk} will be unsafe.}

\item{min_count}{a count lower than \code{min_count} will be unsafe.}
}
\value{
\link[raster:RasterLayer]{raster::RasterLayer} with unsafe cells set to \code{NA}.
}
\description{
remove unsafe cells from raster
}
\examples{
library(raster)
library(magrittr)

b <- block_estimate(dwellings, "unemployed", r=1e3)

# plot the normally rastered data
b \%>\%
  plot(zlim=c(0,1))

b \%>\%
  remove_unsafe() \%>\%
  plot(zlim=c(0,1))

}
\seealso{
Other disclosed: \code{\link{disclosure_risk}},
  \code{\link{is_unsafe}}
}
\concept{disclosed}
