% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcSpatial-package.R
\docType{package}
\name{sdcSpatial-package}
\alias{sdcSpatial}
\alias{sdcSpatial-package}
\title{Privacy Protected maps}
\description{
Privacy protected raster maps 
  can be created from spatial point data. Protection
  methods include smoothing, suppressing revealing values and quadtreeing.
}
\references{
de Jonge, E., & de Wolf, P. P. (2016, September). Spatial smoothing and statistical disclosure control. In International Conference on Privacy in Statistical Databases (pp. 107-117). Springer, Cham.

de Wolf, P. P., & de Jonge, E. (2018, September). Safely Plotting Continuous Variables on a Map. In International Conference on Privacy in Statistical Databases (pp. 347-359). Springer, Cham.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwindj/sdcSpatial}
  \item Report bugs at \url{https://github.com/edwindj/sdcSpatial/issues}
}

}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (0000-0002-6580-4718)

Authors:
\itemize{
  \item Peter-Paul de Wolf
}

Other contributors:
\itemize{
  \item Sapphire Han [contributor]
}

}
