% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_score.R
\name{sensitivity_score}
\alias{sensitivity_score}
\title{mean sensitivity score for raster}
\usage{
sensitivity_score(x, max_risk = x$max_risk, min_count = x$min_count,
  ...)
}
\arguments{
\item{x}{\link{sdc_raster} object created with \code{\link[=sdc_raster]{sdc_raster()}}}

\item{max_risk}{a risk value higher than \code{max_risk} will be sensitive.}

\item{min_count}{a count lower than \code{min_count} will be sensitive.}

\item{...}{passed on to \code{\link{is_sensitive}}}
}
\description{
The mean sensitity score calculates the fraction of cells (with a value)
that are considered sensitive according to the used \code{\link{disclosure_risk}}
}
