% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_rel_freq.R, R/smooth.R
\name{smooth}
\alias{smooth}
\title{Utility function for smoothing}
\usage{
smooth(x, bw = raster::res(x), na.rm = TRUE, pad = TRUE,
  threshold = 10, ...)

smooth(x, bw = raster::res(x), na.rm = TRUE, pad = TRUE,
  threshold = 10, ...)
}
\arguments{
\item{x}{raster object}

\item{bw}{bandwidth}

\item{na.rm}{should the \code{NA} value be removed from the raster?}

\item{pad}{should the data be padded.}

\item{threshold}{cells with a lower (weighted) value of this threshold will be removed.}

\item{...}{passed through to focal}
}
\description{
Approximates a kde by applying a gaussian blur and rescaling.

Create kde density version of a raster
}
