% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdown_code.R
\name{topdown_code}
\alias{topdown_code}
\title{Remove revealing high and low values}
\usage{
topdown_code(r, top = Inf, down = -Inf, ..., upper = top,
  lower = down)
}
\description{
Remove revealing high and low values
}
\keyword{internal}
