% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raster.R
\name{create_raster}
\alias{create_raster}
\title{Create a raster at a certain resolution}
\usage{
create_raster(x, res = 200, ...)
}
\arguments{
\item{x}{geographical object (that supports extent).}

\item{res}{desired resolution (default 200).}

\item{...}{passed through to \code{\link[raster:raster]{raster::raster()}} function.}
}
\value{
\code{\link[raster:raster]{raster::raster()}} object
}
\description{
Utility function to generate a raster at a certain resolution (with rounded coordinates).
}
\keyword{internal}
