% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{disclosure_risk}
\alias{disclosure_risk}
\title{Calculate disclosure risk}
\usage{
disclosure_risk(r, type = c("external", "internal", "discrete"))
}
\arguments{
\item{r}{\link[raster:brick]{raster::brick} object created with \code{\link[=calc_block]{calc_block()}}}

\item{type}{what kind of measure should be used.}
}
\value{
\link[raster:RasterLayer]{raster::RasterLayer} object with the disclosure risk.
}
\description{
Calculate disclosure risk
}
\seealso{
Other disclosed: \code{\link{is_unsafe}},
  \code{\link{remove_unsafe}}
}
\concept{disclosed}
