% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_risk.R
\name{disclosure_risk}
\alias{disclosure_risk}
\title{Calculate disclosure risk}
\usage{
disclosure_risk(x, type = c("external", "internal", "discrete"))
}
\arguments{
\item{x}{\link{sdc_raster} object created with \code{\link[=sdc_raster]{sdc_raster()}}}

\item{type}{what kind of measure should be used.}
}
\value{
\link[raster:raster]{raster::raster} object with the disclosure risk.
}
\description{
Calculate disclosure risk
}
\seealso{
Other disclosed: \code{\link{is_sensitive}},
  \code{\link{remove_sensitive}}
}
\concept{disclosed}
