% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dwellings}
\alias{dwellings}
\title{Data set generated with residence locations  from the open data \href{https://zakelijk.kadaster.nl/bag-producten}{BAG register}
The locations are realistic, but the associated data is simulated.}
\format{An object of class \code{data.frame} with 90603 rows and 2 columns.}
\usage{
dwellings
}
\description{
\itemize{
\item \code{sens_cont} \code{numeric}, simulated sensitive continuous value
\item \code{sens_discrete} \code{logical}, simulated discrete value
}
}
\references{
Basisregistratie Adressen en Gebouwen https://zakelijk.kadaster.nl/bag-producten
}
\keyword{datasets}
