% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwellings.R
\docType{data}
\name{dwellings}
\alias{dwellings}
\title{Simulated dwellings data set}
\format{An object of class \code{data.frame} with 90603 rows and 4 columns.}
\usage{
dwellings
}
\description{
The data are generated with residence locations from the Dutch open data \href{https://zakelijk.kadaster.nl/bag-producten}{BAG register}
The locations are realistic, but the associated data is simulated.
}
\details{
\itemize{
\item \code{x} \code{integer}, x coordinate of dwelling
\item \code{y} \code{integer}, y coordinate of dwelling
\item \code{consumption} \code{numeric}, simulated continuous value
\item \code{unemployed} \code{logical}, simulated discrete value
}
}
\references{
Basisregistratie Adressen en Gebouwen https://zakelijk.kadaster.nl/bag-producten
}
\keyword{datasets}
