% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sdc_raster}
\alias{plot.sdc_raster}
\title{Plot a sdc_raster object}
\usage{
\method{plot}{sdc_raster}(x, fun = mean, ...,
  main = paste(substitute(x)))
}
\arguments{
\item{x}{\link{sdc_raster} object to be plotted}

\item{fun}{function to be used to derive values, e.g. \link{sum} or \link{mean} (default).}

\item{...}{passed on to \code{\link[raster:plot]{raster::plot()}}}

\item{main}{title of plot}
}
\description{
Plot a sdc_raster object
}
\seealso{
Other plotting: \code{\link{plot_sensitive}}
}
\concept{plotting}
