% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sdc_raster}
\alias{plot.sdc_raster}
\title{Plot a sdc_raster object}
\usage{
\method{plot}{sdc_raster}(
  x,
  value = "mean",
  sensitive = TRUE,
  ...,
  main = paste(substitute(x)),
  col
)
}
\arguments{
\item{x}{\code{\link{sdc_raster}} object to be plotted}

\item{value}{\code{character} which value layer to be used for plotting, e.g. "sum", "count", "mean" (default).}

\item{sensitive}{\code{logical} show the sensitivity in the plot?}

\item{...}{passed on to \code{\link[raster:plot]{raster::plot()}}}

\item{main}{title of plot}

\item{col}{color palette to be used, passed on to \code{\link[raster:plot]{raster::plot()}}.}
}
\description{
Plot a sdc_raster object together with its sensitivity.
}
\details{
When \code{sensitive} is set to \code{TRUE}, a side-by-side plot will be made of
the \code{value} and its \code{sensitivity}.
}
\seealso{
Other plotting: 
\code{\link{plot_sensitive}()}
}
\concept{plotting}
