% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sensitive}
\alias{plot_sensitive}
\title{Plot the sensitive cells of the sdc_raster.}
\usage{
plot_sensitive(x, value = "mean", main = "sensitive", col, ...)
}
\arguments{
\item{x}{\link{sdc_raster} object}

\item{value}{\link{character} which value layer to be used for values,
e.g. "sum", "count", "mean" (default).}

\item{main}{\link{character} title of map.}

\item{col}{color palette to be used, passed on to \code{\link[raster:plot]{raster::plot()}}.}

\item{...}{passed on to \code{\link{plot.sdc_raster}}.}
}
\description{
Plots t the sensitive cells of the sdc_raster. The sensitive cells are
plotted in red. The sensitive cells are determined using \code{\link{is_sensitive}}.
}
\seealso{
Other plotting: 
\code{\link{plot.sdc_raster}()}

Other sensitive: 
\code{\link{disclosure_risk}()},
\code{\link{is_sensitive}()},
\code{\link{is_sensitive_at}()},
\code{\link{remove_sensitive}()},
\code{\link{sdc_raster}()},
\code{\link{sensitivity_score}()}
}
\concept{plotting}
\concept{sensitive}
