% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_quadtree.R
\name{protect_quadtree}
\alias{protect_quadtree}
\title{Protect a raster with  a quadtree method.}
\usage{
protect_quadtree(x, max_zoom = Inf, ...)
}
\arguments{
\item{x}{\code{\link{sdc_raster}} object}

\item{max_zoom}{\code{numeric}, restricts the number of zoom and thus the max resolution.
The max resolution = resolution * 2^max_zoom.}

\item{...}{Arguments passed on to \code{is_sensitive}
\describe{
  \item{x}{\link{sdc_raster} object created with \code{\link[=sdc_raster]{sdc_raster()}}}
  \item{max_risk}{a risk value higher than \code{max_risk} will be sensitive.}
  \item{min_count}{a count lower than \code{min_count} will be sensitive.}
  \item{risk_type}{what kind of measure should be used. (Details)}
}}
}
\value{
a \code{\link{sdc_raster}} object, in which sensitive cells have been recursively aggregated until not sensitive.
}
\description{
\code{protect_quadtree} reduces sensitivy by aggregating sensisitve cells with its three neighbors, and
does this recursively until no sensitive cells are left.
}
\details{
This implementation generalizes the method as described by Suñé et al., in which there is no
risk function, and only a  \code{min_count} to determine sensitivity.
Furthermore the method the article
only handles count data (\code{x$value$count}), not mean or summed values.
Currently the translation feature of the article is not (yet) implemented, for the
original method does not take the \code{disclosure_risk} into account.
}
\examples{
library(raster)

fined <- sdc_raster(enterprises, enterprises$fined)
plot(fined)
fined_qt <- protect_quadtree(fined)
plot(fined_qt)

fined <- sdc_raster(enterprises, enterprises$fined, r=50)
plot(fined)
fined_qt <- protect_quadtree(fined)
plot(fined_qt, col=hcl.colors(20, "Blues", rev=TRUE))

}
\references{
Suñé, E., Rovira, C., Ibáñez, D., Farré, M. (2017). Statistical disclosure control on visualising geocoded population data using
a structure in quadtrees, NTTS 2017
}
\seealso{
Other protection methods: \code{\link{protect_smooth}},
  \code{\link{remove_sensitive}}
}
\concept{protection methods}
