% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_smooth.R
\name{protect_smooth}
\alias{protect_smooth}
\title{Protect a sdc_raster by smoothing}
\usage{
protect_smooth(x, bw = raster::res(x$value), ...)
}
\arguments{
\item{x}{raster object}

\item{bw}{bandwidth}

\item{...}{passed through to \code{\link{focal}}.}
}
\description{
Protect a sdc_raster by smoothing
}
\examples{
library(raster)

production <- sdc_raster(enterprises, "production")
zlim <- c(0, 3e4)

plot(production, zlim=zlim)

production_smoothed <- protect_smooth(production, bw = 400)
plot(production_smoothed, zlim=zlim)

production_smooth_safe  <- remove_unsafe(production_smoothed, min_count=2)
plot(production_smooth_safe, zlim=zlim)
}
