% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_smooth.R
\name{protect_smooth}
\alias{protect_smooth}
\title{Protect a sdc_raster by smoothing}
\usage{
protect_smooth(x, bw = raster::res(x$value), ...)
}
\arguments{
\item{x}{raster object}

\item{bw}{bandwidth}

\item{...}{passed through to \code{\link{focal}}.}
}
\description{
Protect a sdc_raster by smoothing.
}
\examples{
library(sdcSpatial)
library(raster)

# create a sdc_raster from point data with raster with resolution of 200m
production <- sdc_raster(enterprises, variable = "production", r = 200, min_count = 3)

print(production)

# plot the raster
zlim <- c(0, 3e4)
# show which raster cells are sensitive
plot(production, zlim=zlim)

# but we can also retrieve directly the raster
sensitive <- is_sensitive(production, min_count = 3)
plot(sensitive, col = c('white', 'red'))

# what is the sensitivy fraction?
sensitivity_score(production)
# or equally
cellStats(sensitive, mean)

# let's smooth to reduce the sensitivity
smoothed <- protect_smooth(production, bw = 400)
plot(smoothed)

# what is the sensitivy fraction?
sensitivity_score(smoothed)

# let's remove the sensitive data.
smoothed_safe <- remove_sensitive(smoothed, min_count = 3)
plot(smoothed_safe, zlim=zlim)

# let's communicate!
production_mean <- mean(smoothed_safe)
production_total <- sum(smoothed_safe)

# and cread
filledContour(production_mean, nlevels = 6, main = "Mean production")
filledContour(production_total, nlevels = 15, col = hcl.colors(21, rev=TRUE), main="Total production")
}
\seealso{
Other protection methods: \code{\link{protect_quadtree}},
  \code{\link{remove_sensitive}}
}
\concept{protection methods}
