% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protect_smooth.R
\name{protect_smooth}
\alias{protect_smooth}
\title{Protect a sdc_raster by smoothing}
\usage{
protect_smooth(x, bw = raster::res(x$value), ...)
}
\arguments{
\item{x}{raster object}

\item{bw}{bandwidth}

\item{...}{passed through to \code{\link{focal}}.}
}
\description{
Protect a sdc_raster by smoothing
}
\examples{
library(raster)
library(magrittr)

production <- sdc_raster(enterprises, "production")
plot(production)


production \%>\%
  protect_smooth(bw = 400) \%>\%
  plot(zlim=c(0,2e4))

production \%>\%
  protect_smooth(bw = 400) \%>\%
  remove_unsafe(min_count=2) \%>\%
  plot(zlim=c(0,2e4))

production \%>\%
  is_unsafe(min_count=2) \%>\%
  hist()

production \%>\%
  protect_smooth(bw=600) \%>\%
  disclosure_risk() \%>\%
  hist()
}
