% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unsafe.R
\name{remove_unsafe}
\alias{remove_unsafe}
\title{remove unsafe cells from raster}
\usage{
remove_unsafe(r, max_risk = 0.95, min_count = 10, ...)
}
\arguments{
\item{r}{\link[raster:brick]{raster::brick} object created with \code{\link[=calc_block]{calc_block()}}}

\item{max_risk}{a risk value higher than \code{max_risk} will be unsafe.}

\item{min_count}{a count lower than \code{min_count} will be unsafe.}
}
\value{
\link[raster:RasterLayer]{raster::RasterLayer} with unsafe cells set to \code{NA}.
}
\description{
remove unsafe cells from raster
}
\examples{
library(raster)

b <- calc_block_logical(dwellings, "unemployed", r=1e3)

# plot the normally rastered data
plot(b$mean, zlim=c(0,1))

r_s <- remove_unsafe(b, type="discrete")

hist(b$mean, xlim=c(0,1), ylim=c(0,150))
hist(r_s, xlim=c(0,1), ylim=c(0,150))

plot(r_s, zlim=c(0,1))
}
\seealso{
Other disclosed: \code{\link{disclosure_risk}},
  \code{\link{is_unsafe}}
}
\concept{disclosed}
