% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_raster.R
\name{sdc_raster}
\alias{sdc_raster}
\title{Raster map with privacy awareness}
\usage{
sdc_raster(x, variable, r = 200, max_risk = 0.95, min_count = 10,
  risk_type = c("external", "internal", "discrete"), ...,
  field = variable)
}
\arguments{
\item{x}{\link[sp:SpatialPointsDataFrame]{sp::SpatialPointsDataFrame}, \link[sf:sf]{sf::sf} or a two column matrix or \link{data.frame}
that is used to create a raster.}

\item{variable}{name of data column or \code{numeric} with same length as \code{x}
to be used for the data.}

\item{r}{either a desired resolution or a pre-existing \link[raster:raster]{raster::raster} object. In the first case, the
crs of \code{x} (if present) will be used, in the latter the properties of the \code{r} will be kept.}

\item{max_risk}{\code{numeric}, the maximum_risk score (\code{\link{disclosure_risk}}) before the data is considered sensitive}

\item{min_count}{\code{numeric}, cells with a number of observations that are less then \code{min_count} are considered
sensitive}

\item{risk_type}{passed on to \code{\link[=disclosure_risk]{disclosure_risk()}}.}

\item{...}{passed through to \code{\link[raster:rasterize]{raster::rasterize()}}}

\item{field}{synonym for \code{variable}. If both supplied, \code{field} has precedence.}
}
\value{
object of \code{class} "sdc_raster":
\itemize{
\item \code{$value}: \code{\link[raster:brick]{raster::brick()}} object with differenct layers e.g. \code{count}, \code{sum}, \code{mean}.
\item \code{$max_risk}: the threshold at the \code{\link[=disclosure_risk]{disclosure_risk()}} is considered sensitive.
\item \code{$min_count}: the minimum number of observation before a cell is \strong{not} sensitive
\item \code{$scale}: used together with \code{min_count} to determine sensitivity: result
of protection operation \code{\link[=protect_smooth]{protect_smooth()}} or \code{\link[=protect_quadtree]{protect_quadtree()}}.
\item \code{$type}: of \code{variable}, either \code{numeric} or \code{logical}
\item \code{$risk_type}, "external", "internal" or "discrete".
}
}
\description{
\code{sdc_raster} derives \code{\link[raster:raster]{raster::raster}} objects from the supplied point data and calculates
statistics needed for controling their sensitivity to disclosure.
The sensitivity of the data can be shown with \code{\link[=plot_sensitive]{plot_sensitive()}}.
}
\examples{
library(raster)
prod <- sdc_raster(enterprises, field = "production", r = 500)
print(prod)

prod <- sdc_raster(enterprises, field = "production", r = 1e3)
print(prod)

# get raster with the average production per cell averaged over the enterprises
prod_mean <- mean(prod)
summary(prod_mean)

# get raster with the total production per cell
prod_total <- sum(prod)
summary(prod_total)
}
\seealso{
Other sensitive: \code{\link{disclosure_risk}},
  \code{\link{is_sensitive}},
  \code{\link{remove_sensitive}},
  \code{\link{sensitivity_score}}
}
\concept{sensitive}
