% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Create kde density version of a raster}
\usage{
smooth(x, bw = res(x), na.rm = TRUE, pad = TRUE, threshold = 10,
  ...)
}
\arguments{
\item{x}{raster object}

\item{bw}{bandwidth}

\item{na.rm}{should the NA value be removed from the raster?}

\item{...}{passed through to focal}
}
\description{
Create kde density version of a raster
}
